+++ tests/test_remotehandler_s3_transfer.py
@@ -308,6 +308,57 @@ s3_to_s3_pca_rename_task_definition = {
     ],
 }
 
+s3_to_s3_rename_task_definition = {
+    "type": "transfer",
+    "source": {
+        "bucket": BUCKET_NAME,
+        "directory": "src",
+        "fileRegex": "file-rename-(.*)\\.txt",
+        "protocol": {
+            "name": "opentaskpy.addons.aws.remotehandlers.s3.S3Transfer",
+        },
+    },
+    "destination": [
+        {
+            "bucket": BUCKET_NAME_2,
+            "directory": "dest",
+            "rename" : {
+                "pattern": "abc",
+                "sub" : "def"
+            },
+            "protocol": {
+                "name": "opentaskpy.addons.aws.remotehandlers.s3.S3Transfer",
+            },
+        },
+    ],
+}
+
+s3_to_s3_proxy_rename_task_definition = {
+    "type": "transfer",
+    "source": {
+        "bucket": BUCKET_NAME,
+        "directory": "src",
+        "fileRegex": "file-rename-proxy-(.*)\\.txt",
+        "protocol": {
+            "name": "opentaskpy.addons.aws.remotehandlers.s3.S3Transfer",
+        },
+    },
+    "destination": [
+        {
+            "bucket": BUCKET_NAME_2,
+            "directory": "dest",
+            "rename" : {
+                "pattern": "abc",
+                "sub" : "def"
+            },
+            "transferType": "proxy",
+            "protocol": {
+                "name": "opentaskpy.addons.aws.remotehandlers.s3.S3Transfer",
+            },
+        },
+    ],
+}
+
 
 s3_to_ssh_copy_task_definition = {
     "type": "transfer",
@@ -799,6 +850,45 @@ def test_s3_to_s3_copy_pca_rename(setup_bucket, tmp_path, s3_client):
         objects["Contents"][0]["Key"] == "src/archive/file-pca-rename-1234-renamed.txt"
     )
 
+def test_s3_to_s3_copy_rename(setup_bucket, tmp_path, s3_client):
+    transfer_obj = transfer.Transfer(
+        None, "s3-to-s3-rename", s3_to_s3_rename_task_definition
+    )
+
+    # Write a test file locally
+    fs.create_files([{f"{tmp_path}/file-rename-abc.txt": {"content": "test1234"}}])
+    create_s3_file(
+        s3_client,
+        f"{tmp_path}/file-rename-abc.txt",
+        "src/file-rename-abc.txt",
+    )
+
+    assert transfer_obj.run()
+
+    # Check that the file is in the destination bucket with new name
+    objects = s3_client.list_objects(Bucket=BUCKET_NAME_2)
+    assert len(objects["Contents"]) == 1
+    assert objects["Contents"][0]["Key"] == "dest/file-rename-def.txt"
+
+def test_s3_to_s3_proxy_rename(setup_bucket, s3_client, tmp_path):
+    transfer_obj = transfer.Transfer(
+        None, "s3-to-s3-proxy-rename", s3_to_s3_proxy_rename_task_definition
+    )
+
+    # Write a test file locally
+    fs.create_files([{f"{tmp_path}/file-rename-proxy-abc.txt": {"content": "test1234"}}])
+    create_s3_file(
+        s3_client,
+        f"{tmp_path}/file-rename-proxy-abc.txt",
+        "src/file-rename-proxy-abc.txt",
+    )
+
+    assert transfer_obj.run()
+
+    # Check that the file is in the destination bucket with new name
+    objects = s3_client.list_objects(Bucket=BUCKET_NAME_2)
+    assert len(objects["Contents"]) == 1
+    assert objects["Contents"][0]["Key"] == "dest/file-rename-proxy-def.txt"
 
 def test_s3_file_watch_custom_creds(
     setup_bucket,