+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v24.14.1
+
+- Add test coverage for s3 file renames when using proxy transfer type
+
+## v24.14.0
+
+- Add destination file rename to S3 transfers
+
 ## v24.13.0
 
 - Fix temporary token not refreshing correctly